Split web type contracts
